Golden Projects/Voice Assistant using Python/requirements.txt
SpeechRecognition==3.8.1
beautifulsoup4==4.9.3
psutil==5.8.0
PyAutoGUI==0.9.52
pyjokes==0.6.0
PyQt5==5.15.2
PyQt5-Qt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3
pyttsx3==2.90
pywhatkit==4.7
requests==2.31.0
urllib3==1.26.6
wikipedia==1.4.0
DateTime==4.3
opencv-python==4.5.2.52
cv2-tools==2.4.0
secure-smtplib==0.1.1
random2==1.0.1
instaloader==4.7.1
PyPDF2>=1.27.5
bs4==0.0.1
pywikihow==0.5.7
speedtest-cli==2.1.3
pytube==10.8.1
os-sys==2.1.4
subprocess32==3.5.4
qt5-applications
qt5-tools
Pillow>=9.0.1
PyWave==0.5.0
phonenumbers==8.12.23
folium==0.12.1
opencage==1.2.2
ffmpeg-normalize==1.22.1
PyAudio==0.2.13